Preparation to use common functions for text and binary mode
The logic to process incoming requests is largely the same for text and
binary mode, which led to lots of code duplication in the previous
library.

By using function pointers, which point to different serialization/
deserialization functions depending on the ThingSet mode, the code
duplication can be dramatically reduced.

This commit prepares the text mode for re-use of some functions in
the binary mode.